Upgrade MongoDB and move validations to middleware

+++ tests/db-handler.js
@@ -1,3 +1,4 @@
+require('dotenv').config();
 const mongoose = require('mongoose');
 
 const connect = async () => {
@@ -7,8 +8,10 @@ const connect = async () => {
 };
 
 const close = async () => {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
+  if (mongoose.connection.readyState !== 0) {
+    await mongoose.connection.dropDatabase();
+    await mongoose.connection.close();
+  }
 };
 
 const clear = async () => {
